refactor: split resource panel editor to ResourceEditor
