+++ src/page/Editor/components/ActionEditor/ActionEditorPanel/ResourceEditor/EventHandler/index.tsx
@@ -1,6 +1,6 @@
-import { handlerTitleStyle } from "../style"
-import { EventInstance } from "./EventInstance"
 import { useTranslation } from "react-i18next"
+import { handlerTitleStyle } from "@/page/Editor/components/ActionEditor/ActionEditorPanel/style"
+import { EventInstance } from "./EventInstance"
 
 export const EventHandler = () => {
   const { t } = useTranslation()
