+++ src/page/Editor/components/ActionEditor/ActionEditorPanel/ResourceEditor/EventHandler/interface.tsx
