+++ src/page/Editor/components/ActionEditor/ActionEditorPanel/ResourceEditor/ResourcePanel/interface.ts
