+++ src/page/Editor/components/ActionEditor/ActionEditorPanel/ResourceEditor/ResourceParams/index.tsx
