+++ src/page/Editor/components/ActionEditor/ActionEditorPanel/ResourceEditor/ResourceParams/interface.tsx
