+++ src/page/Editor/components/ActionEditor/ActionEditorPanel/ResourceEditor/index.tsx
@@ -1,0 +1,119 @@
+import { FC, useContext, useRef } from "react"
+import { css } from "@emotion/react"
+import { useTranslation } from "react-i18next"
+import { useSelector } from "react-redux"
+import { Select, Option } from "@illa-design/select"
+import { PenIcon } from "@illa-design/icon"
+import { Divider } from "@illa-design/divider"
+import { selectAllResource } from "@/redux/currentApp/action/resource/resourceSelector"
+import { ResourcePanel } from "@/page/Editor/components/ActionEditor/ActionEditorPanel/ResourceEditor/ResourcePanel"
+import { ActionEditorContext } from "@/page/Editor/components/ActionEditor/context"
+import { triggerRunRef } from "@/page/Editor/components/ActionEditor/ActionEditorPanel/interface"
+import {
+  actionStyle,
+  fillingStyle,
+  actionSelectStyle,
+  triggerSelectStyle,
+  resourceSelectStyle,
+  applyEditIconStyle,
+  sectionTitleStyle,
+  resourceBarStyle,
+  resourceOptionStyle,
+  resourceBarTitleStyle,
+  panelScrollStyle,
+} from "@/page/Editor/components/ActionEditor/ActionEditorPanel/style"
+import { ResourceEditorProps } from "./interface"
+
+export const ResoucreEditor: FC<ResourceEditorProps> = (props) => {
+  const {
+    onChangeResource,
+    onCreateResource,
+    onEditResource,
+    onChangeParam,
+    onSaveParam,
+  } = props
+
+  const { t } = useTranslation()
+  const { resourceId } = useContext(ActionEditorContext)
+  const resourceList = useSelector(selectAllResource)
+  const triggerRunRef = useRef<triggerRunRef>(null)
+
+  const triggerOptions = [
+    {
+      label: t("editor.action.panel.option.trigger.manually"),
+      value: 0,
+    },
+    {
+      label: t("editor.action.panel.option.trigger.on_change"),
+      value: 1,
+    },
+  ]
+
+  const isResourceEditable = resourceId && resourceId.indexOf("preset") === -1
+
+  return (
+    <div css={panelScrollStyle}>
+      <div css={css(actionStyle, resourceBarStyle)}>
+        <label css={css(sectionTitleStyle, resourceBarTitleStyle)}>
+          {t("editor.action.panel.label.resource")}
+        </label>
+        <span css={fillingStyle} />
+        <Select
+          options={triggerOptions}
+          defaultValue={0}
+          css={css(actionSelectStyle, triggerSelectStyle)}
+        />
+
+        <Select
+          css={css(actionSelectStyle, resourceSelectStyle)}
+          value={resourceId}
+          onChange={onChangeResource}
+          triggerProps={{
+            autoAlignPopupWidth: false,
+          }}
+        >
+          <Option onClick={onCreateResource} isSelectOption={false}>
+            <span
+              css={resourceOptionStyle}
+              title={t("editor.action.panel.option.resource.new")}
+            >
+              {t("editor.action.panel.option.resource.new")}
+            </span>
+          </Option>
+          <Divider />
+          <Option value={"preset_REST API"}>
+            <span
+              css={resourceOptionStyle}
+              title={t("editor.action.panel.option.resource.rest_query")}
+            >
+              {t("editor.action.panel.option.resource.rest_query")}
+            </span>
+          </Option>
+          {resourceList &&
+            resourceList.map(({ resourceId: id, resourceName: name }) => (
+              <Option value={id} key={id}>
+                <span css={resourceOptionStyle} title={name}>
+                  {name}
+                </span>
+              </Option>
+            ))}
+        </Select>
+        <div
+          css={applyEditIconStyle(!isResourceEditable)}
+          onClick={onEditResource}
+        >
+          <PenIcon />
+        </div>
+      </div>
+      <Divider />
+      <ResourcePanel
+        ref={triggerRunRef}
+        resourceId={resourceId}
+        onChange={onChangeParam}
+        onSave={onSaveParam}
+      />
+    </div>
+  )
+}
+
+ResoucreEditor.displayName = "ResourceEditor"
