+++ src/page/Editor/components/ActionEditor/ActionEditorPanel/ResourceEditor/interface.tsx
@@ -1,0 +1,7 @@
+export interface ResourceEditorProps {
+  onChangeResource?: (resourceId: string) => void
+  onCreateResource?: () => void
+  onEditResource?: () => void
+  onChangeParam?: () => void
+  onSaveParam?: () => void
+}
