+++ src/page/Editor/components/ActionEditor/ActionEditorPanel/index.tsx
@@ -1,41 +1,27 @@
 import { FC, useMemo, useState, useRef, useContext } from "react"
-import { css } from "@emotion/react"
 import { v4 as uuidV4 } from "uuid"
 import { Button } from "@illa-design/button"
-import { Select, Option } from "@illa-design/select"
-import { Divider } from "@illa-design/divider"
-import { CaretRightIcon, MoreIcon, PenIcon } from "@illa-design/icon"
+import { CaretRightIcon, MoreIcon } from "@illa-design/icon"
 import { Dropdown } from "@illa-design/dropdown"
 import { Menu } from "@illa-design/menu"
 import { useTranslation } from "react-i18next"
 import { useDispatch, useSelector } from "react-redux"
 import { selectAllActionItem } from "@/redux/currentApp/action/actionList/actionListSelector"
-import { selectAllResource } from "@/redux/currentApp/action/resource/resourceSelector"
 import { actionListActions } from "@/redux/currentApp/action/actionList/actionListSlice"
 import { ActionEditorContext } from "@/page/Editor/components/ActionEditor/context"
 import { generateName } from "@/page/Editor/components/ActionEditor/utils"
+import { ResoucreEditor } from "@/page/Editor/components/ActionEditor/ActionEditorPanel/ResourceEditor"
+import { TitleInput } from "@/page/Editor/components/ActionEditor/ActionEditorPanel/TitleInput"
 import { ActionEditorPanelProps, triggerRunRef } from "./interface"
 import {
   containerStyle,
   headerStyle,
-  actionStyle,
   fillingStyle,
-  actionSelectStyle,
-  triggerSelectStyle,
-  resourceSelectStyle,
-  applyEditIconStyle,
   moreBtnStyle,
-  sectionTitleStyle,
-  resourceBarStyle,
-  panelScrollStyle,
   moreBtnMenuStyle,
   duplicateActionStyle,
   deleteActionStyle,
-  resourceOptionStyle,
-  resourceBarTitleStyle,
 } from "./style"
-import { TitleInput } from "./TitleInput"
-import { ResourcePanel } from "./ResourcePanel"
 import { ActionResult } from "./ActionResult"
 
 const { Item: MenuItem } = Menu
@@ -59,7 +45,6 @@ export const ActionEditorPanel: FC<ActionEditorPanelProps> = (props) => {
   const [moreBtnMenuVisible, setMoreBtnMenuVisible] = useState(false)
   const triggerRunRef = useRef<triggerRunRef>(null)
   const actionItems = useSelector(selectAllActionItem)
-  const resourceList = useSelector(selectAllResource)
   const activeActionItem = useMemo(() => {
     if (!activeActionItemId) {
       return null
@@ -72,17 +57,6 @@ export const ActionEditorPanel: FC<ActionEditorPanelProps> = (props) => {
   }, [actionItems])
 
   const isResourceEditable = resourceId && resourceId.indexOf("preset") === -1
-
-  const triggerOptions = [
-    {
-      label: t("editor.action.panel.option.trigger.manually"),
-      value: 0,
-    },
-    {
-      label: t("editor.action.panel.option.trigger.on_change"),
-      value: 1,
-    },
-  ]
 
   function createResource() {
     onCreateResource && onCreateResource()
@@ -184,72 +158,18 @@ export const ActionEditorPanel: FC<ActionEditorPanelProps> = (props) => {
             : t("editor.action.panel.btn.run")}
         </Button>
       </header>
-      <div css={panelScrollStyle}>
-        {activeActionItem && (
-          <>
-            <div css={css(actionStyle, resourceBarStyle)}>
-              <label css={css(sectionTitleStyle, resourceBarTitleStyle)}>
-                {t("editor.action.panel.label.resource")}
-              </label>
-              <span css={fillingStyle} />
-              <Select
-                options={triggerOptions}
-                defaultValue={0}
-                css={css(actionSelectStyle, triggerSelectStyle)}
-              />
-
-              <Select
-                css={css(actionSelectStyle, resourceSelectStyle)}
-                value={resourceId}
-                onChange={onChangeResource}
-                triggerProps={{
-                  autoAlignPopupWidth: false,
-                }}
-              >
-                <Option onClick={createResource} isSelectOption={false}>
-                  <span
-                    css={resourceOptionStyle}
-                    title={t("editor.action.panel.option.resource.new")}
-                  >
-                    {t("editor.action.panel.option.resource.new")}
-                  </span>
-                </Option>
-                <Divider />
-                <Option value={"preset_REST API"}>
-                  <span
-                    css={resourceOptionStyle}
-                    title={t("editor.action.panel.option.resource.rest_query")}
-                  >
-                    {t("editor.action.panel.option.resource.rest_query")}
-                  </span>
-                </Option>
-                {resourceList &&
-                  resourceList.map(({ resourceId: id, resourceName: name }) => (
-                    <Option value={id} key={id}>
-                      <span css={resourceOptionStyle} title={name}>
-                        {name}
-                      </span>
-                    </Option>
-                  ))}
-              </Select>
-              <div
-                css={applyEditIconStyle(!isResourceEditable)}
-                onClick={editResource}
-              >
-                <PenIcon />
-              </div>
-            </div>
-            <Divider />
-            <ResourcePanel
-              ref={triggerRunRef}
-              resourceId={resourceId}
-              onChange={onChange}
-              onSave={onSave}
-            />
-          </>
-        )}
-      </div>
-      <ActionResult />
+      {activeActionItem && (
+        <>
+          <ResoucreEditor
+            onChangeParam={onChange}
+            onSaveParam={onSave}
+            onCreateResource={createResource}
+            onEditResource={editResource}
+            onChangeResource={onChangeResource}
+          />
+          <ActionResult />
+        </>
+      )}
     </div>
   )
 }
