+++ src/page/Editor/components/ActionEditor/Resource/RESTAPI/param/index.tsx
@@ -3,7 +3,7 @@ import { css } from "@emotion/react"
 import { useTranslation } from "react-i18next"
 import { Select } from "@illa-design/select"
 import { Input } from "@illa-design/input"
-import { FieldArray } from "@/page/Editor/components/ActionEditor/ActionEditorPanel/FieldArray"
+import { FieldArray } from "@/page/Editor/components/ActionEditor/ActionEditorPanel/ResourceEditor/FieldArray"
 import { ActionEditorContext } from "@/page/Editor/components/ActionEditor/context"
 import { useSelector } from "react-redux"
 import { selectAllActionItem } from "@/redux/currentApp/action/actionList/actionListSelector"
